% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stb.R
\name{stb}
\alias{stb}
\title{Generic Method for computing Simultaneous Tolerance Bounds (bands and intervals) 
for fitted models or numeric vectors.}
\usage{
stb(obj, ...)
}
\arguments{
\item{obj}{(object) passed on}

\item{...}{additional parameters}
}
\description{
If the generic method is applied to an object of class \code{VCA} function \code{\link{stb.VCA}}
is called and all its arguments may be specified. Otherwise, function \code{\link{stb.default}}
is called and all its arguments may be specified. The latter was developed for numeric vectors
(see ?stb.default). Therefore this method will most likely produce strange results on other
types of objects than \code{VCA} and numeric vectors.
}

